requirements.txt
asgiref
Django
django-cors-headers
django-filter
djangorestframework
djangorestframework-simplejwt
PyJWT==1.7.1
pytz
sqlparse
typing-extensions
gunicorn
whitenoise
djangorestframework-api-key
docutils
psycopg2
sendgrid
